+++ examples/basic.rs
@@ -117,13 +117,13 @@ impl Action for MyCustomAction {
                     lerp_type: LerpType::Rotation(Quat::from_look(Vec3::Z, Vec3::Y)),
                     duration: 1.0,
                 }),
-                WaitAction::new(1.0).into_boxed(),
+                WaitAction::new(1.0),
                 LerpAction::new(LerpConfig {
                     target: camera,
                     lerp_type: LerpType::Position(CAMERA_OFFSET),
                     duration: 1.0,
                 }),
-                WaitAction::new(0.5).into_boxed(),
+                WaitAction::new(0.5),
                 LerpAction::new(LerpConfig {
                     target: agent,
                     lerp_type: LerpType::Rotation(Quat::from_look(-Vec3::Z, Vec3::Y)),