+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [Add linked actions][63]
 - [Replace `config` with `start`, `order` and `repeat` methods][64]
+- [Replace `IntoBoxedAction` trait with `From<Box<dyn Action>>`][65]
 
 ## Version 0.6.0
 
@@ -43,6 +44,7 @@
 
 First release! 🎉
 
+[65]: https://github.com/hikikones/bevy-sequential-actions/pull/65
 [64]: https://github.com/hikikones/bevy-sequential-actions/pull/64
 [63]: https://github.com/hikikones/bevy-sequential-actions/pull/63
 [55]: https://github.com/hikikones/bevy-sequential-actions/pull/55